Fehler korrigiert

+++ source/PflanzenPI.Persistence/Schema/Model/BrightnessEntity.cs
@@ -1,3 +1,3 @@
 ﻿namespace PflanzenPI.Persistence.Schema;
 
-public record BrightnessEntity(string owner, int hourLux);
+public record BrightnessEntity(string owner, double hourLux);
